+++ src/components/presentational/index.js
@@ -1,0 +1,19 @@
+import Login from './Login'
+import Logo from './Logo'
+import MainTop from './MainTop'
+import Member from './Member'
+import MobileNav from './MobileNav'
+import NotFound from './NotFound'
+import Owners from './Owners'
+import UserAvatar from './UserAvatar'
+
+export {
+  Login,
+  Logo,
+  MainTop,
+  Member,
+  MobileNav,
+  NotFound,
+  Owners,
+  UserAvatar
+}
